refactor(architecture): simplify config and ports
