+++ internal/adapters/logger.go
@@ -2,20 +2,22 @@ package adapters
 
 import (
 	"log/slog"
-
-	"github.com/ekalinin/github-markdown-toc.go/v2/internal/core/ports"
 )
+
+type logger interface {
+	Info(string, ...any)
+}
 
 type Logger struct {
 	debug bool
-	log   ports.Logger
+	log   logger
 }
 
 func NewLogger(debug bool) *Logger {
 	return NewLoggerX(debug, slog.Default())
 }
 
-func NewLoggerX(debug bool, logger ports.Logger) *Logger {
+func NewLoggerX(debug bool, logger logger) *Logger {
 	return &Logger{
 		debug: debug,
 		log:   logger,
