+++ internal/adapters/remoteposter.go
@@ -3,24 +3,10 @@ package adapters
 import (
 	"context"
 	"net/http"
-
-	"github.com/ekalinin/github-markdown-toc.go/v2/internal/core/ports"
-	"github.com/ekalinin/github-markdown-toc.go/v2/internal/utils"
 )
 
-type realPoster struct {
-	client *http.Client
-}
-
-func (p *realPoster) Post(ctx context.Context, url, token, path string) (string, error) {
-	if err := ctx.Err(); err != nil {
-		return "", err
-	}
-	return utils.HttpPost(ctx, p.client, url, path, token)
-}
-
 type RemotePoster struct {
-	poster ports.RemotePoster
+	client *http.Client
 }
 
 func NewRemotePoster() *RemotePoster {
@@ -28,13 +14,12 @@ func NewRemotePoster() *RemotePoster {
 }
 
 func NewRemotePosterWithClient(client *http.Client) *RemotePoster {
-	return NewRemotePosterX(&realPoster{client: client})
-}
-
-func NewRemotePosterX(poster ports.RemotePoster) *RemotePoster {
-	return &RemotePoster{poster: poster}
+	return &RemotePoster{client: client}
 }
 
 func (r *RemotePoster) Post(ctx context.Context, url, token, path string) (string, error) {
-	return r.poster.Post(ctx, url, token, path)
+	if err := ctx.Err(); err != nil {
+		return "", err
+	}
+	return HttpPost(ctx, r.client, url, path, token)
 }
